Unify Context Inputs and Inputs
