+++ schema-tool/schemas/input.py
@@ -40,3 +40,82 @@ def UTXOInput(
 
 
 AVAILABLE_SCHEMAS.append(UTXOInput())
+
+# Commitment Input
+
+commitment_input_name = "Commitment Input"
+commitment_input_fields: List[Field] = [
+    input_type_field(1, commitment_input_name),
+    SimpleField(
+        "Commitment ID", ByteArray(36), "The commitment identifier to reference to."
+    ),
+]
+
+
+def CommitmentInput(
+    omitFields: bool = False,
+) -> Schema:
+    return Schema(
+        commitment_input_name,
+        "A Commitment Input allows referencing a commitment to a certain slot and is used to provide a notion of time for transaction execution that is linked to the containing Block's <code>Issuing Time</code>.",
+        commitment_input_fields,
+        tipReference=45,
+        omitFields=omitFields,
+    )
+
+
+AVAILABLE_SCHEMAS.append(CommitmentInput())
+
+# Block Issuance Credit Input
+
+block_issuance_credit_input_name = "Block Issuance Credit Input"
+block_issuance_credit_input_fields: List[Field] = [
+    input_type_field(2, block_issuance_credit_input_name),
+    SimpleField(
+        "Account ID",
+        ByteArray(32),
+        "The ID of the Account for which this input provides the BIC.",
+    ),
+]
+
+
+def BlockIssuanceCreditInput(
+    omitFields: bool = False,
+) -> Schema:
+    return Schema(
+        block_issuance_credit_input_name,
+        "A Block Issuance Credit Input provides the BIC balance of a specific account as context to transaction execution.",
+        block_issuance_credit_input_fields,
+        tipReference=45,
+        omitFields=omitFields,
+    )
+
+
+AVAILABLE_SCHEMAS.append(BlockIssuanceCreditInput())
+
+# Reward Input
+
+reward_input_name = "Reward Input"
+reward_input_fields: List[Field] = [
+    input_type_field(3, reward_input_name),
+    SimpleField(
+        "Index",
+        UInt16(),
+        "The index of the transaction input for which to claim rewards.",
+    ),
+]
+
+
+def RewardInput(
+    omitFields: bool = False,
+) -> Schema:
+    return Schema(
+        reward_input_name,
+        "A Reward Input indicates which transaction Input is claiming Mana rewards.",
+        reward_input_fields,
+        tipReference=45,
+        omitFields=omitFields,
+    )
+
+
+AVAILABLE_SCHEMAS.append(RewardInput())
